refactor: split api system handlers into focused modules
